# swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Me Socorre API",
    "description": "Documentação da API da plataforma de prestação de serviços Me Socorre.",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Ambiente local"
    }
  ],
...
  "paths": {
...
    "/admin/dashboard": {
      "get": {
        "tags": [
          "Administração"
        ],
        "summary": "Consulta os indicadores administrativos",
        "description": "Retorna totais e agrupamentos usados no painel administrativo.",
        "responses": {
          "200": {
            "description": "Indicadores retornados com sucesso."
          },
          "401": {
            "description": "Token não informado, inválido ou expirado."
          },
          "403": {
            "description": "Acesso permitido apenas para administradores."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/admin/providers": {
      "get": {
        "tags": [
          "Administração"
        ],
        "summary": "Lista os prestadores para administração",
        "description": "Lista todos os prestadores, com filtro opcional pelo status de aprovação.",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "enum": {
                  "type": "array",
                  "example": [
                    "PENDING",
                    "APPROVED",
                    "REJECTED",
                    "SUSPENDED",
                    "BANNED"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              }
            },
            "description": "Status de aprovação do prestador."
          }
        ],
        "responses": {
          "200": {
            "description": "Prestadores retornados com sucesso."
          },
          "400": {
            "description": "Filtro de status inválido."
          },
          "401": {
            "description": "Token não informado, inválido ou expirado."
          },
          "403": {
            "description": "Acesso permitido apenas para administradores."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/admin/providers/{id}/approve": {
      "patch": {
        "tags": [
          "Administração"
        ],
        "summary": "Aprova um prestador",
        "description": "Aprova um prestador que está com o cadastro pendente.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "format": {
                  "type": "string",
                  "example": "uuid"
                }
              }
            },
            "description": "ID do perfil do prestador."
          }
        ],
        "responses": {
          "200": {
            "description": "Prestador aprovado com sucesso."
          },
          "400": {
            "description": "ID inválido."
          },
          "401": {
            "description": "Token não informado, inválido ou expirado."
          },
          "403": {
            "description": "Acesso permitido apenas para administradores."
          },
          "404": {
            "description": "Prestador não encontrado."
          },
          "409": {
            "description": "Transição de status não permitida."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/admin/providers/{id}/reject": {
      "patch": {
        "tags": [
          "Administração"
        ],
        "summary": "Rejeita um prestador",
        "description": "Rejeita um prestador que está com o cadastro pendente.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "format": {
                  "type": "string",
                  "example": "uuid"
                }
              }
            },
            "description": "ID do perfil do prestador."
          }
        ],
        "responses": {
          "200": {
            "description": "Prestador rejeitado com sucesso."
          },
          "400": {
            "description": "ID inválido."
          },
          "401": {
            "description": "Token não informado, inválido ou expirado."
          },
          "403": {
            "description": "Acesso permitido apenas para administradores."
          },
          "404": {
            "description": "Prestador não encontrado."
          },
          "409": {
            "description": "Transição de status não permitida."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/admin/providers/{id}/suspend": {
      "patch": {
        "tags": [
          "Administração"
        ],
        "summary": "Suspende um prestador",
        "description": "Suspende um prestador que está aprovado.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "format": {
                  "type": "string",
                  "example": "uuid"
                }
              }
            },
            "description": "ID do perfil do prestador."
          }
        ],
        "responses": {
          "200": {
            "description": "Prestador suspenso com sucesso."
          },
          "400": {
            "description": "ID inválido."
          },
          "401": {
            "description": "Token não informado, inválido ou expirado."
          },
          "403": {
            "description": "Acesso permitido apenas para administradores."
          },
          "404": {
            "description": "Prestador não encontrado."
          },
          "409": {
            "description": "Transição de status não permitida."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/admin/providers/{id}/ban": {
      "patch": {
        "tags": [
          "Administração"
        ],
        "summary": "Bane um prestador",
        "description": "Bane um prestador pendente, aprovado ou suspenso.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "format": {
                  "type": "string",
                  "example": "uuid"
                }
              }
            },
            "description": "ID do perfil do prestador."
          }
        ],
        "responses": {
          "200": {
            "description": "Prestador banido com sucesso."
          },
          "400": {
            "description": "ID inválido."
          },
          "401": {
            "description": "Token não informado, inválido ou expirado."
          },
          "403": {
            "description": "Acesso permitido apenas para administradores."
          },
          "404": {
            "description": "Prestador não encontrado."
          },
          "409": {
            "description": "Transição de status não permitida."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/admin/providers/{id}/featured": {
      "patch": {
        "tags": [
          "Administração"
        ],
        "summary": "Altera o destaque de um prestador",
        "description": "Adiciona ou remove um prestador aprovado da lista de destaques.",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "string"
                },
                "format": {
                  "type": "string",
                  "example": "uuid"
                }
              }
            },
            "description": "ID do perfil do prestador."
          }
        ],
        "responses": {
          "200": {
            "description": "Destaque atualizado com sucesso."
          },
          "400": {
            "description": "ID ou dados inválidos."
          },
          "401": {
            "description": "Token não informado, inválido ou expirado."
          },
          "403": {
            "description": "Acesso permitido apenas para administradores."
          },
          "404": {
            "description": "Prestador não encontrado."
          },
          "409": {
            "description": "Prestador ainda não foi aprovado."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "destaque"
                ],
                "properties": {
                  "destaque": {
                    "type": "boolean",
                    "example": true
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/reviews": {
      "get": {
        "tags": [
          "Administração"
        ],
        "summary": "Lista todas as avaliações",
        "description": "Retorna todas as avaliações para acompanhamento administrativo.",
        "responses": {
          "200": {
            "description": "Avaliações retornadas com sucesso."
          },
          "401": {
            "description": "Token não informado, inválido ou expirado."
          },
          "403": {
            "description": "Acesso permitido apenas para administradores."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/admin/requests": {
      "get": {
        "tags": [
          "Administração"
        ],
        "summary": "Lista todas as solicitações",
        "description": "Retorna todas as solicitações para acompanhamento administrativo.",
        "responses": {
          "200": {
            "description": "Solicitações retornadas com sucesso."
          },
          "401": {
            "description": "Token não informado, inválido ou expirado."
          },
          "403": {
            "description": "Acesso permitido apenas para administradores."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/admin/clients": {
      "get": {
        "tags": [
          "Administração"
        ],
        "summary": "Lista todos os clientes",
        "description": "Retorna todos os clientes e as quantidades de solicitações e avaliações.",
        "responses": {
          "200": {
            "description": "Clientes retornados com sucesso."
          },
          "401": {
            "description": "Token não informado, inválido ou expirado."
          },
          "403": {
            "description": "Acesso permitido apenas para administradores."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ]
      }
    },
    "/reviews/": {
      "post": {
        "tags": [
          "Avaliações"
        ],
        "summary": "Avalia um atendimento",
        "description": "Permite que o cliente avalie uma solicitação concluída uma única vez.",
        "responses": {
          "201": {
            "description": "Avaliação criada com sucesso."
          },
          "400": {
            "description": "Dados inválidos."
          },
          "401": {
            "description": "Token não informado, inválido ou expirado."
          },
          "403": {
            "description": "Cliente sem permissão para avaliar a solicitação."
          },
          "404": {
            "description": "Solicitação não encontrada."
          },
          "409": {
            "description": "Solicitação não concluída ou já avaliada."
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "idSolicitacao",
                  "nota"
                ],
                "properties": {
                  "idSolicitacao": {
                    "type": "integer",
                    "minimum": 1,
                    "example": 15
                  },
                  "nota": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5,
                    "example": 5
                  },
                  "comentario": {
                    "type": "string",
                    "minLength": 3,
                    "maxLength": 1000,
                    "example": "Atendimento rápido e eficiente."
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/providers/": {
      "get": {
        "tags": [
          "Prestadores"
        ],
        "summary": "Lista os prestadores disponíveis",
        "description": "Retorna prestadores aprovados e disponíveis, com filtros opcionais.",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            },
            "description": "Busca por nome, categoria ou endereço."
          },
          {
            "name": "categoryId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "integer"
                },
                "minimum": {
                  "type": "number",
                  "example": 1
                }
              }
            },
            "description": "Filtra pelo ID da categoria."
          }
        ],
        "responses": {
          "200": {
            "description": "Prestadores retornados com sucesso."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Informe o token JWT retornado pelo endpoint de login."
      }
    }
  }
}